+++ common/src/main/java/io/github/jamalam360/reaping/PlayerEntityDuck.java
@@ -1,0 +1,11 @@
+package io.github.jamalam360.reaping;
+
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.ActionResult;
+
+/**
+ * @author Jamalam360
+ */
+public interface PlayerEntityDuck {
+    ActionResult reapingmod$onReaped(ItemStack toolStack);
+}