+++ common/src/main/java/io/github/jamalam360/reaping/config/ReapingConfig.java
@@ -12,5 +12,7 @@ public interface ReapingConfig {
 
     boolean reapBabies();
 
+    boolean reapPlayers();
+
     int deathChance();
 }