+++ fabric/src/main/java/io/github/jamalam360/reaping/fabric/ReapingExpectPlatformImpl.java
@@ -6,7 +6,10 @@ import io.github.jamalam360.reaping.fabriclike.mixin.LootContextBuilderAccessor;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.damage.DamageSource;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.loot.context.LootContext;
+import virtuoel.pehkui.api.ScaleData;
+import virtuoel.pehkui.api.ScaleTypes;
 
 public class ReapingExpectPlatformImpl {
     public static ReapingConfig getConfig() {
@@ -17,7 +20,12 @@ public class ReapingExpectPlatformImpl {
         return FabricLoader.getInstance().isModLoaded(modId);
     }
 
-    public static  LootContext.Builder getLootContextBuilder(LivingEntity entity, boolean causedByPlayer, DamageSource source) {
+    public static LootContext.Builder getLootContextBuilder(LivingEntity entity, boolean causedByPlayer, DamageSource source) {
         return ((LootContextBuilderAccessor) entity).callGetLootContextBuilder(causedByPlayer, source);
+    }
+
+    public static void setScale(PlayerEntity player, float scale) {
+        ScaleData data = ScaleTypes.BASE.getScaleData(player);
+        data.setTargetScale(scale);
     }
 }