+++ quilt/src/main/java/io/github/jamalam360/reaping/fabric/ReapingExpectPlatformImpl.java
@@ -5,8 +5,11 @@ import io.github.jamalam360.reaping.fabriclike.ReapingFabricLike;
 import io.github.jamalam360.reaping.fabriclike.mixin.LootContextBuilderAccessor;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.damage.DamageSource;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.loot.context.LootContext;
 import org.quiltmc.loader.api.QuiltLoader;
+import virtuoel.pehkui.api.ScaleData;
+import virtuoel.pehkui.api.ScaleTypes;
 
 /**
  * @author Jamalam360
@@ -20,7 +23,12 @@ public class ReapingExpectPlatformImpl {
         return QuiltLoader.isModLoaded(modId);
     }
 
-    public static  LootContext.Builder getLootContextBuilder(LivingEntity entity, boolean causedByPlayer, DamageSource source) {
+    public static LootContext.Builder getLootContextBuilder(LivingEntity entity, boolean causedByPlayer, DamageSource source) {
         return ((LootContextBuilderAccessor) entity).callGetLootContextBuilder(causedByPlayer, source);
+    }
+
+    public static void setScale(PlayerEntity player, float scale) {
+        ScaleData data = ScaleTypes.BASE.getScaleData(player);
+        data.setTargetScale(scale);
     }
 }