+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 - Add a configurable chance of death when an animal is reaped.
+- Allow reaping of players, which causes the player to shrink dramatically.
 
 ### Fixes